requirements.txt
openai>=1.0.0
python-telegram-bot==20.7
python-dotenv==1.0.1
langdetect==1.0.9
